requirements.txt
# Core CLI and Framework
typer>=0.9.0
rich>=13.0.0
pydantic>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
click>=8.0.0

# Async and HTTP
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
requests>=2.31.0

# Security and Encryption
cryptography>=41.0.0

# AI Providers
anthropic>=0.25.0
openai>=1.0.0

# Social Media APIs
tweepy>=4.14.0
atproto>=0.0.40  # Bluesky AT Protocol
Mastodon.py>=1.8.0  # Mastodon API
praw>=7.0.0  # Reddit API Wrapper
google-api-python-client>=2.0.0  # YouTube Data API
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Development and Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0